Update setup.sh
trying to improve setup.sh

modified this file so that a user can remotely execute this without downloading it first

rclone mv command was failing so I modified it to cp -rf. cp will overwrite files in a directory and makes the whole script more robust and it can be used as a "updater" as well.

Rclone command now pulls latest stable.

added "git" to the install list

added command to auto-move git repo contents to users root directory.

hardcoded rclone directory's because the asterisks in or with the variables were failing for me.

Please let me know if any of this is helpful.